Splitup code into seperate codefiles
